requirements.txt
# Core Libraries (lightweight)
numpy>=1.24.0
pillow>=9.0.0

# FastAPI (Backend)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.5

# HTTP requests for online model loading
requests>=2.28.0

# Logging
loguru>=0.7.0

# Testing Dependencies
pytest>=7.0.0
pytest-mock>=3.10.0
httpx>=0.24.0